code is refactored
