--- public/js/whiteboard.js
@@ -1,10 +1,0 @@
-$('.container').ready(function (){
-  var $unchosen = $('.unchosen').draggable({
-    containment:  $('.row'),
-    cursor     : 'pointer',
-    snap       :$('.grouping')
-  }
-);
-
-
-});//closes document.ready
